Refactor: move color-related code to separate modules
