+++ docs/docs/palettes/app/color/generate.js
@@ -1,0 +1,3 @@
+export { generateGrays, generateGrays as grays } from './generate-grays.js';
+export { generatePalette, generatePalette as palette } from './generate-palette.js';
+export { generateScale, generateScale as scale } from './generate-scale.js';
